# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  go-micro_internal_payment.CreatePaymentRequest:
    properties:
      amount:
        type: integer
      order_id:
        type: string
      request_id:
        type: string
    type: object
...
  go-micro_pkg_httpx.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_gateway.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_gateway.Item'
        type: array
      remark:
        type: string
      request_id:
        type: string
      user_id:
        type: string
    type: object
  internal_gateway.Item:
    properties:
      price:
        type: integer
      quantity:
        type: integer
      sku_id:
        type: string
    type: object
...
  internal_inventory.ConfirmRequest:
    properties:
      reserved_id:
        type: string
    type: object
  internal_inventory.Item:
    properties:
      quantity:
        type: integer
      sku_id:
        type: string
    type: object
  internal_inventory.ReleaseByOrderRequest:
    properties:
      order_id:
        type: string
    type: object
  internal_inventory.ReleaseRequest:
    properties:
      reserved_id:
        type: string
    type: object
  internal_inventory.ReserveRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_inventory.Item'
        type: array
      order_id:
        type: string
    type: object
  internal_order.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_order.Item'
        type: array
      remark:
        type: string
      request_id:
        type: string
      user_id:
        type: string
    type: object
  internal_order.Item:
    properties:
      price:
        type: integer
      quantity:
        type: integer
      sku_id:
        type: string
    type: object
  internal_payment.CreatePaymentRequest:
    properties:
      amount:
        type: integer
      order_id:
        type: string
      request_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Go-Micro 网关服务 API
  title: Go-Micro Gateway API
  version: "1.0"
paths:
...
  /api/v1/orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 订单创建请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_gateway.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建订单
      tags:
      - Order
  /api/v1/orders/{id}:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 订单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询订单
      tags:
      - Order
  /api/v1/payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建支付请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/go-micro_internal_payment.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建支付
      tags:
      - Payment
  /api/v1/payments/{id}:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询支付
      tags:
      - Payment
  /api/v1/payments/{id}/failed:
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付失败
      tags:
      - Payment
  /api/v1/payments/{id}/success:
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付成功
      tags:
      - Payment
  /api/v1/payments/{id}/timeout:
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付超时
      tags:
      - Payment
...
  /inventory/{sku_id}:
    get:
      parameters:
      - description: SKU ID
        in: path
        name: sku_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询库存
      tags:
      - Inventory
  /inventory/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: 确认请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_inventory.ConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 确认扣减
      tags:
      - Inventory
  /inventory/release:
    post:
      consumes:
      - application/json
      description: 基于 reserved_id 释放，重复调用幂等
      parameters:
      - description: 释放请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_inventory.ReleaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 释放库存
      tags:
      - Inventory
  /inventory/release-by-order:
    post:
      consumes:
      - application/json
      description: 基于订单号幂等释放库存，重复调用不会重复回补
      parameters:
      - description: 订单号释放请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_inventory.ReleaseByOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 按订单号释放
      tags:
      - Inventory
  /inventory/reservations/{order_id}:
    get:
      parameters:
      - description: 订单ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询预占记录
      tags:
      - Inventory
  /inventory/reserve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 预占请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_inventory.ReserveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 预占库存
      tags:
      - Inventory
  /orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建订单请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_order.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建订单
      tags:
      - Order
  /orders/{id}:
    get:
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询订单
      tags:
      - Order
  /orders/{id}/cancel:
    post:
      description: 幂等语义：已取消返回成功，已完成不可取消
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 取消订单
      tags:
      - Order
  /orders/biz/{biz_no}:
    get:
      parameters:
      - description: 业务订单号
        in: path
        name: biz_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 按业务单号查询
      tags:
      - Order
  /payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建支付请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_payment.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 创建支付
      tags:
      - Payment
  /payments/{id}:
    get:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 查询支付
      tags:
      - Payment
  /payments/{id}/failed:
    post:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付失败
      tags:
      - Payment
  /payments/{id}/refund:
    post:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 退款
      tags:
      - Payment
  /payments/{id}/success:
    post:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付成功
      tags:
      - Payment
  /payments/{id}/timeout:
    post:
      parameters:
      - description: 支付ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/go-micro_pkg_httpx.Response'
      summary: 支付超时
      tags:
      - Payment
...
swagger: "2.0"